cvi_mpi: weekly rls 2025.08.15
68d35b, Merge "[feat](auido):v410 support musl32
e2adc2, Merge "[fix]v410 support musl32" into v4.1.0
196c11, Merge "[fix](bin): add musl32 built libcvi_json-c.a and libcvi_miniz.a

+++ modules/isp/cv181x/isp/src/isp_dpc_ctrl.c
@@ -392,7 +392,7 @@ static CVI_S32 dump_raw(VI_PIPE ViPipe, CVI_U16 *bayerBuffer,
 			ViPipe, order_id,
 			width,
 			height,
-			16, tv1.tv_sec, tv1.tv_usec
+			16, (long int)tv1.tv_sec, (long int)tv1.tv_usec
 			);
 
 	output = fopen(img_name, "wb");
